feat: CompareGraph 리팩토링
