+++ frontend/sucpi/src/api.js
@@ -1,5 +1,8 @@
 const API_URL = process.env.REACT_APP_API_URL || "http://localhost";
 
+/**
+ * 리더보드 데이터를 가져오는 함수
+ */
 export const fetchLeaderboardData = async () => {
     try {
         const response = await fetch(`${API_URL}/api/admin/leaderboard`, {
@@ -20,6 +23,10 @@ export const fetchLeaderboardData = async () => {
     }
 };
 
+/**
+ * 학생 데이터를 가져오는 함수
+ * @param {string} id - 학생 ID
+ */
 export const fetchStudentData = async (id) => {
     const response = await fetch(`${API_URL}/api/students/${id}`);
     if (!response.ok) {
@@ -27,4 +34,49 @@ export const fetchStudentData = async (id) => {
     }
     const data = await response.json(); // JSON 데이터를 변수에 저장
     return data.result; // 저장된 데이터를 반환
+};
+
+/**
+ * 가중치 데이터를 가져오는 함수
+ */
+export const fetchWeights = async () => {
+    const response = await fetch(`${API_URL}/api/admin/weights`);
+    if (!response.ok) {
+        throw new Error('가중치를 가져오는 데 실패했습니다.');
+    }
+    const data = await response.json();
+    return data.result; // `result`만 반환
+};
+
+/**
+ * 특정 학생의 그래프 데이터를 가져오는 함수
+ * @param {string} studentId - 학생 ID
+ */
+export const fetchGraphData = async (studentId) => {
+    const response = await fetch(`${API_URL}/api/admin/weights/test/${studentId}`);
+    if (!response.ok) {
+        throw new Error('그래프 데이터를 가져오는 데 실패했습니다.');
+    }
+    return response.json();
+};
+
+/**
+ * 비교 데이터를 서버로 전송하는 함수
+ * @param {string} studentId - 학생 ID
+ * @param {object} updatedData - 비교 데이터
+ */
+export const postComparisonData = async (studentId, updatedData) => {
+    console.log("JSON -> ", JSON.stringify(updatedData));
+    const response = await fetch(`${API_URL}/api/admin/weights/test/${studentId}`, {
+        method: 'POST',
+        headers: {
+            'Content-Type': 'application/json',
+        },
+        body: JSON.stringify(updatedData),
+    });
+
+    if (!response.ok) {
+        throw new Error('비교 데이터를 서버로 전송하는 데 실패했습니다.');
+    }
+    return response.json();
 };
